requirements.txt
asgiref==3.2.3
attrs==19.3.0
Django==3.0.2
djangorestframework==3.11.0
importlib-metadata==1.4.0
more-itertools==8.1.0
packaging==20.1
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.6
pytest==5.3.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
wcwidth==0.1.8
zipp==2.1.0
pytest-django

